Update to latest version of BICv3 for RC
(Hopefully all the things work better)

+++ index.php
@@ -38,13 +38,28 @@ if (array_key_exists('themeSwatch', $content) && $content['themeSwatch']){
         <title>BICv3</title>
         <meta name="viewport" content="width=device-width, initial-scale=1">
         <?php
-        if(array_key_exists('themePack', $content) && strlen($content['themePack']) > 0){
+        if (array_key_exists('themePack', $content) && strlen($content['themePack']) > 0) {
             echo '<link rel="stylesheet" href="' . $cdna->getURI($asConfig['themePack']) . '" />';
         } else {
-            echo '<link rel="stylesheet" href="https://d1c6dfkb81l78v.cloudfront.net/jquery.mobile/1.3.0/jqm.theme.min.css" />';
+            if (array_key_exists('platformCSS', $content) && strlen($content['platformCSS']) > 0 && $content['platformCSS'] === 'default') {
+                echo '<link rel="stylesheet" href="https://d1c6dfkb81l78v.cloudfront.net/jquery.mobile/1.3.0/jqm.theme.min.css" />';
+            }
+        }
+
+        if (array_key_exists('platformCSS', $content) && strlen($content['platformCSS']) > 0 && $content['platformCSS'] !== 'none') {
+            echo '<link rel="stylesheet" href="https://d1c6dfkb81l78v.cloudfront.net/jquery.mobile/1.3.0/jqm.structure.min.css" />';
         }
         ?>
-        <link rel="stylesheet" href="https://d1c6dfkb81l78v.cloudfront.net/jquery.mobile/1.3.0/jqm.structure.min.css" />
+       
+        <script>
+          <?php
+          if (strpos($_SERVER[HTTP_USER_AGENT], 'BlinkGap') === FALSE){
+            echo 'window.NativeApp = false;';
+          } else {
+            echo 'window.NativeApp = true;';
+          }
+          ?>
+        </script>
         <script data-main="/_BICv3_/source/main" src="https://d1c6dfkb81l78v.cloudfront.net/requirejs/2.1.2/require.min.js"></script>
         <?php
         if(array_key_exists('externalJavaScript', $content) && strlen($content['externalJavaScript']) > 0){
@@ -66,8 +81,8 @@ if (array_key_exists('themeSwatch', $content) && $content['themeSwatch']){
         }
         ?>
     </head>
-        <body>
-            <noscript>You currently have JavaScript disabled. This application requires JavaScript to work correctly.</noscript>
-            <div data-role="page" id="temp">Loading, please wait.</div>
-        </body>
-    </html>
+    <body>
+        <noscript>You currently have JavaScript disabled. This application requires JavaScript to work correctly.</noscript>
+        <div data-role="page" id="temp">Loading, please wait.</div>
+    </body>
+</html>
